Move item container code into 'interactive'
