+++ src/game_logic/ingame_systems.hpp
@@ -41,8 +41,8 @@
 #include "game_logic/enemies/spike_ball.hpp"
 #include "game_logic/interactive/elevator.hpp"
 #include "game_logic/interactive/enemy_radar.hpp"
+#include "game_logic/interactive/item_container.hpp"
 #include "game_logic/interactive/sliding_door.hpp"
-#include "game_logic/item_container.hpp"
 #include "game_logic/player.hpp"
 #include "game_logic/player/damage_system.hpp"
 #include "game_logic/player/projectile_system.hpp"
